+++ src/controllers/DefaultController.php
@@ -9,10 +9,14 @@
 namespace ether\gnash\controllers;
 
 use Craft;
+use craft\db\Table;
+use craft\errors\MissingComponentException;
 use craft\web\Controller;
 use ether\gnash\Gnash;
 use yii\db\Exception;
+use yii\db\Query;
 use yii\web\BadRequestHttpException;
+use yii\web\Response;
 
 /**
  * Class DefaultController
@@ -24,16 +28,26 @@ class DefaultController extends Controller
 {
 
 	/**
+	 * @return Response
+	 * @throws BadRequestHttpException
 	 * @throws Exception
+	 * @throws MissingComponentException
 	 */
 	public function actionPurgeAll ()
 	{
 		Gnash::getInstance()->gnash->purgeAll();
+
+		Craft::$app->getSession()->setNotice(
+			Craft::t('nginx-cache', 'Cache Cleared')
+		);
+
+		return $this->redirectToPostedUrl();
 	}
 
 	/**
 	 * @throws BadRequestHttpException
 	 * @throws Exception
+	 * @throws MissingComponentException
 	 */
 	public function actionPurgeElement ()
 	{
@@ -41,19 +55,36 @@ class DefaultController extends Controller
 
 		$elementType = $request->getRequiredBodyParam('elementType');
 		$elementIds = $request->getRequiredBodyParam('element-' . $elementType);
+		$relatedTo = (bool) $request->getBodyParam('relatedTo', false);
+
+		if ($relatedTo)
+			$elementIds = Gnash::getInstance()->gnash->getRelatedIds($elementIds);
 
 		foreach ($elementIds as $id)
-			Gnash::getInstance()->gnash->purgeElement($id);
+			Gnash::getInstance()->gnash->purgeElement((int) $id);
+
+		Craft::$app->getSession()->setNotice(
+			Craft::t('nginx-cache', 'Cache Cleared')
+		);
+
+		return $this->redirectToPostedUrl();
 	}
 
 	/**
 	 * @throws BadRequestHttpException
 	 * @throws Exception
+	 * @throws MissingComponentException
 	 */
 	public function actionPurgeUrl()
 	{
 		$url = Craft::$app->getRequest()->getRequiredBodyParam('url');
 		Gnash::getInstance()->gnash->purgeUrl($url);
+
+		Craft::$app->getSession()->setNotice(
+			Craft::t('nginx-cache', 'Cache Cleared')
+		);
+
+		return $this->redirectToPostedUrl();
 	}
 
 }